requirements.txt
Flask==3.1.1
psycopg2-binary==2.9.10
redis==6.4.0